CLI: Remove IP whitelist
Although IP whitelisting is a great feature, it's easier to implement at
the network level by opening and closing the entire port that cimpler is
listening on instead. Also, it was checking the current server's IP
address rather than the remote address so you'd have to add the current
server's IP to the list which allows access to from everywhere.

This just removes the whitelist entirely to simplify configuration.