Tests: use spawn() in cli test
It's safer now that we're passing an uncontrolled path to a config file
on the cli. The saftey comes from passing args as an array of strings
which can be used in the system call, rather than have bash interpret
them.

Also, in further commits, we'll be adding more params. An array is
nicer / safer to deal with than just hoping they don't contain
characters that would need escaping.