git-build: wait for event loop after the proc exits
Previously, some tests would fail with "write after end".

The finishBuild() function closes the logfile. The proc's stdout /
stderr is being piped to the logfile and we can't interrupt or flush it.
It's possible that there are still bytes in the pipe waiting to be
written to the log. Let's wait a turn of the event loop before closing
the logFile.