Post additional build status message on `buildAdded`
This carries the risk that we will add status messages to pulls which will never be built and will stay in the queue forever, due to the `buildFilter` configuration. I don't think that's an incorrect behavior; I think we should probably have a separate system to dismiss such stranded pulls from the queue, which could then be hooked in here.